+++ test/fixtures/composite_event_transforming_card.js
@@ -1,3 +1,6 @@
+// TODO: Update test environment to not depend on this for older browsers.
+Conductor.Oasis.configure('allowSameOrigin', true);
+
 Conductor.card({
   activate: function () {
     ok(true, "Card was activated");