+++ History.md
@@ -1,5 +1,7 @@
 ### 0.2.0
 
+- `Conductor.configure('eventCallback', function (eventHandler) {})` to configure
+  a wrapper around message event handlers.
 - cards may defer their activation by returning a promise from `activate`.
 - custom services will have `error` invoked if the environment did not provide
   the capability.  For example: